requirements.txt
python-telegram-bot==21.5
apscheduler==3.10.4
requests==2.31.0
groq==0.9.0
python-dotenv==1.0.1
feedparser==6.0.11